Add MagnetStingAdvanced class to magnetsting.py
Created a new class called "MagnetStingAdvanced". This class is similar
to the "MagnetSting" class, however, it offers additional capabilities.
It now gives the user to create commands that take a command prefix and
an argument.

The argument(s) can either be specified in a tuple
(referred to as "strict-type" commands, accepting arguments only
specified in the tuple) or the argument can be anything
(referred to as "free-type" commands). Users can also create "single-type"
commands, which are just command names, operating like the commands in the
"MagnetSting" class.

The module documentation was also updated, adding more information about
the "MagnetSting" class and adding information about the new
"MagnetStingAdvanced" class.