Add additional info to "add_command_single" method
The documentation originally stated that "... a single-command is just
that, a single-command name with no extra stuff.". However, "additional
stuff" could be assigned to the function via the "additional_data"
parameter, so the method documentation was changed to reflect that. It
now reads "... a single-command name with no extra stuff, unless the
`additional_data` param is used, of course.".