Add new parameters to "MagnetStingAdvanced" class
Added the parameters "verbose" and "type_spacer" to the class
initialization method in the "MagnetStingAdvanced" class.

The "verbose" parameter is used to indicate whether or not the help
banner will provide more in depth information. If it is set to "True",
then the help banner will contain the command names, description, as well
as what type of command they are (strict, free, single). If the verbose
parameter is set to "False", then the help banner will only contain the
command names and the descriptions.

The "type_spacer" parameter is used for formatting the spacing between
the command descriptions and the command type columns in the help
banner. This parameter is only used if the "verbose" parameter is set to
"True".