Fix bug that caused "help" and "quit" commands to not show
A bug that caused the "help" and "quit" commands to not appear in the
help banner when the "verbose" parameter in "MagnetStingAdvanced" was
set to "True" has been fixed. The bug was due to the "zip" function not
finding a pair for the two commands in the command type dict, which
caused it to not show them in the help banner. This was alleviated by
assigning the two commands the command type "built-in".