Fix strict-type command options not showing and formatting
When calling "help <strict-type command name>", the options for that
specific command were not showing up due to the elif statement that
handles it being indented wrong. However, the correct indentation has
been implemented and the command options are now shown when calling the
"help" command.

Additionally, the formatting in which the options are listed has also
been changed.