Add "clear" built-in command to "possible commands"
The built-in "clear" command was not appearing when the list of
"possible commands" when MAGNETSTING ran into a command it did not
recognize due to the command not being in the "self._commands_dict"
dictionary. It has now been added and shows in the list of "possible
commands" when applicable.