Remove strict-type commands and begin move to single dict
The strict-type commands have now been removed. The reasoning behind it
is that the same functionality could be implemented in the functions
themselves that are assigned to command names, giving the user more control
over how a function behaves if it gets an option it does not support.

Additionally, the structure of how commands and their associated
information are stored is being changed. Rather than have several
dictionaries for different values of a command, a single dict will now
hold all command-specific information. The command name will serve as
the key, and the value will be another dictionary that will hold the
command-specific information (help message, function, etc.).