Rework documentation
The documentation for the module and the classes has been rewritten.
Whereas before the documentation for both classes was written at the
start of the module, it has now been split based on the classes. The
documentation for the classes was also updated to make it more concise
and understandable and to account for the changes, especially for
MagnetStingAdvanced.