Create new method in MagnetStingAdvanced
The new "_possible_commands" method in MagnetStingAdvanced now handles
showing possible commands based on the user input if the input was not
found in the commands_info dict. Whereas this was all handled in an
"else" statement in "magnetstingadvanced_mainloop", it is now its own
separate method.

Additionally, the way that the possible commands are displayed has been
updated as well. Instead of showing the possible commands in a single
string separated by two spaces, it now prints them in rows of four,
with the output resembling that of the "column" command on UNIX/Linux
systems.

Furthermore, the parameter "command_hint_spacers" has been removed from
the class initialization, as it was not being used and would no longer
be used either, given the changes made.