Rework banner creation in MagnetStingAdvanced
The way in which banners are created has been reworked. A single
parameter ("banner") now handles all aspects of banner creation. It can
determine whether to use the class-generated banner or to use a custom
banner. In additions, the parameters "banner_decorators",
"decorator_length" and "custom_banner" have been removed as a result of
the changes. The "banner_decorators" and "decorator_length" parameters
were removed to give more control over the look of the opening banner.
Now, banner decorators must be added into the "banner" parameter along
with the other things that appear in the banner. The "custom_banner"
parameter was removed because the "banner" parameter now also handles
custom banners.