Rework command menu spacing allocation in MagnetStingAdvanced
The amount of spacing between commands, their help descriptions and the
commands types in the command menu has been changed to automatically
calculate the needed amount of spacing between the columns, rather than
have the user supply the values for the spacing. As a result of the
changes, the parameters "help_spacer" and "type_spacer" have been
removed MagnetStingAdvanced.