Create "_specific_commands_help" method in MagnetStingAdvanced
The "_specific_commands_help" method in MagnetStingAdvanced allows a
user to print the help banner with specific commands by typing "help
<command name>". It does not have to be the entire command name either,
it can just be the first few letters. Rather than have to print the
entire help banner and sift through the various commands looking for the
one that you want, you can just type the name or part of the name and
get a more refined result.