Rework command handling in MagnetStingAdvanced
The way that the built-in commands are handled in MagnetStingAdvanced
has been modified. Rather that use the raw user input, it now uses the
first element in the processed user input list. This change was made so
that the processing of commands is not a mix of checking between both
the raw user input and the processed list, and just relies on checking
the first element in the list instead.