Create built-in command "alias"
The "alias" command now allows for the creation, removal and viewing of
command aliases. The user can create aliases for any command they want,
as long as the command exists. Aliases can be set by using "alias add
<alias name> <command name>", removed using "alias remove <command name>"
and viewed by using "alias view". The aliases act just like shell
aliases and can have additional arguments passed to them. There is also
a safeguard in place that prevents a user from creating aliases with the
same name as a command, so as to avoid any unwanted problems.

Additionally, the way commands are processed in the
"magnetstingadvanced_mainloop" has been modified to account for aliases.