Change list to check for args for free-type commands
Switched the list for length checks for free-type commands from
"split_command" to "full_command_list". This change was due to the "[!]
Argument required" message being shown if an aliased free-type command
was not given any arguments after the alias name, even if the full
command had args. This occurred because the "split_command" list did not
contain the full command that was aliased, rather, it only contained the
alias name and as a result, when the len of "split_command" was checked,
it showed as only 1, causing the error message to be shown. However,
with "full_command_list" being checked instead, it no longer has any
problems dealing with aliased commands.