Modify specific help banner additional spacing calculation
The amount of additional spacing added to the len of the longest command
has been changed from a static amount of 5, to either 5 or 12. If the
len of the longest or only command is less than 5, then an additional 12
is added. If the len is longer than 5, than an additional 5 is then
added. This change was made because when, for example, calling "help q",
the help banner columns and data became misaligned because the command
name was too short, even with the additional 5 spacing. With the changes
however, the the columns and data now remain aligned, regardless of the
len.