Add ability to change JSON alias file name
The ".alias.json" file can now be given any name a user wants. This
change is to avoid problems that arise when there are two or more
instances of MagnetStingAdvanced, in particular when it comes to
reading, adding, editing and removing command aliases.

For example, suppose there are two instances of MagnetStingAdvanced,
"MagnetStingAdvanced_A" and "MagnetStingAdvanced_B", where
"MagnetStingAdvanced_B" is called by command from "MagnetStingAdvanced_A".
Because they would both access ".alias.json", if any commands were
aliased in "MagnetStingAdvanced_A", they would also show up in the
"MagnetStingAdvanced_B" instance, even if the command that was aliased
does not exist in "MagnetStingAdvanced_B". Similarily, if there were
changes made to aliases in "MagnetStingAdvanced_B", they would be loaded
in "MagnetStingAdvanced_A" the next time it is run.

By allowing the name of the JSON file to be changed by using the "alias_file"
paramter in the class initialization of MagnetStingAdvanced, command
aliases can be kept to specific instances and avoids one instance from
altering the aliases of other instances.