Add ability to create "command groups"
The ability to create "command groups" in MagnetStingAdvanced has been
added through the "add_command_group" method. A "command group", as the
name suggests, is a group of commands. Command groups can be used to
organize similar commands, as well as to avoid creating help banners of
excessive lengths that could overwhelm a user.

To add commands to a group, a group first has to be created using the
"add_command_group" method. Any type of command can be added to a
command group, and any command added to a command group will not show up
in the standard help banner, rather, only the group name will. To
actually add a command to a created group, the three command creation
methods have each received a new parameter called "command_group", where
one can enter the name of the command group they wish to assign the
command to.

To see all the commands within a specific group, all one has to do is enter
the command group's name, and it will display a help banner of command
assigned to the group. The syntax to call a command within a group is
<command group name> <command name> <args (if free or parser type)>. Just
like with other commands, commands in command groups can be aliased.

As a result of the addition of the "add_command_group" method, the
command processing internals in the "magnetstingadvanced_mainloop"
methods have been modified to account for command groups.