Add KeyError handling in command groups
When attempting to call a command that did not exist in a command
group, a KeyError was raised because there was no handling in place for
non-existent commands. Now however, instead of a KeyError being raised,
the "_possible_commands" method is called which has been modified to
look through either the "_commands_info" dict or the "_groups_dict".

Additionally, white spacing is now also stripped from the user input.