MagnetStingAdvanced now becomes MagnetSting
The original MagnetSting class has been completly removed.
The reason for the removal is simply because it was not being
maintained and there is no reason to keep an older version of the
framework in the module. MagnetStingAdvanced offers greater
capabilities than its predecessor and all development efforts were
focused on it, with MagnetSting being updated once every blue moon,
if at all. Thus, the decision was made to remove the original
MagnetSting and rename MagnetStingAdvanced to MagnetSting.

Additionally, some of the documentation in the new MagnetSting (formerly
MagnetStingAdvanced) has been updated.