Make slight changes to some formatted strings
Made minor change to some of the formatted strings in terms of how the
code looks, specifically, adding a space between the string and the
"{:spacing}" spacers, as some had space between the string and the
spacer while others did not. This change does not affect the way
anything is printed, this was more for making the formatted strings
follow the same format all throughout the code.