Add new parameter to class initialization
Added the "help_on_start" parameter to the class initialization. The
parameter allows one to choose if they want the help banner to be shown
at the start. Setting the parameter to "True" will display the help
banner on start and setting it to "False" will not. If the parameter is
set to "False", the help banner and the help functionality can still be
called using the "help" command. Documentation for the new parameter was
also added.