Edit documentation of class and methods
The documentation regarding parser-type commands has been edited, both
in the class documentation as well as the method documentation. It now
states that any Python file can be assigned, rather than only
"argparse" files. Moreover, the explanation of the command was also
modified to make it more understandable.

The documentation for the other commands has been slightly modified as
well, such as describing what the method does as well as the structure
of the documentation to account for the changes.

Additionally, any remaining mentions of "any" when it comes to what data
type the "additional_data" parameter can accept has been changed to
"tuple".