Reformat formatted strings
Removed unneccessary spacing from the ends of various formatted strings
as they served no purpose.

Additionally, indentation in all strings that have it is now provided
using a static two spaces, rather than using " "*2.

Moreover, the structure of some of the formatted strings was also
changed, but this does not affect how it is displayed. The changes were
made to make the code more readable.