Fix commands in command groups not working
When attempting to call a single- or file-type command that was in a
command group, a "keyError" exception was raised due to the wrong
dictionary being checked, specifically "self._commands_info". With the
changes, the correct dictionary, "commands_dict", is now used and allows
all types of commands in a command group to be called without any
errors.