Add error handling for aliased non-existent commands
When creating an alias using MagnetSting, it checks to see if the
command being aliased acutally exists, preventing an alias for any
command that does not exist from actually being created. However, there
is nothing preventing someone from directly accessing the alias file
(default ".alias.json") and adding in an alias for a non-existent
command. Or perhaps an alias was created for a command that once existed
but has now been removed for some reason. Either way, the alias still
exists, and when trying to call it, it led to an exception being raised.
Now however, when attempting to call an alias for a non-existent
command, a message is displayed informing the user that the command that
is aliased does not exist.