Add error handling for aliased commands
When attempting to execute an aliased group command, MagnetSting will
now check if the command actually exists within the group. If it does,
it moves to the next part of command processing. If the command does
not exist, it displays a message indicating that the alias cannot be
executed because the command does not exist within the group, rather
than have a "KeyError" exception raised.