Small edit for documentation about args-type commands
When explaining the required parameters for a function that is to be
assigned to an args-type command, the "command_args" parameter was
described as taking a "str" object. However, MagnetSting actually passes
the arguments to the function in the form of a "list". The documentation
was updated, now showing that the parameter takes a list.