Add additional functionality to "help" command
The "help" command has recieved a new functionality. A user can now seek
additional help for a specific command by typing "help <command name>"
(the entire command name doesn't need to be entered, rather, just enough of
it for MagnetSting to understand which command you are looking for). This
allows for a more in-depth explanation of a command which can be
supplied through the new "long_help" parameter that was added to all
command and command group creation methods.