Add exception handling for JSONDecodeError
Added exception handling for JSONDecodeError that may arise if
MagnetSting attempts to decode a malformed JSON object from the alias
file. Rather than cause a crash, an error message is now displayed
informing the user that the aliases could not be loaded due to a
decoding error and that they should check on the alias file.