Update alias file after every addition and removal
The alias file is now updated after every new alias addition/edit and
removal. The reason for this is because MagnetSting used to only update
the alias after exiting. However, should MagnetSting have encountered an
error and crashed, then any changes made to alias dictionary would not
be written to the alias file. Now, with every addition, edit and
removal, the alias file is now updated as well.