Fix IndexError caused when calling aliased group names
When an alias consisting of only a command group name was called, it
resulted in an IndexError being raised. This was due to a lack of proper
handling for such cases when parsing commands. This issue has been
resolved and calling such an alias behaves the same way was as calling
the command group name itself.