Add confirmation before removing all aliases
When using the "alias remove *" command, a confirmation has been added
prior to actually removing all of the aliases to double check with the
user if they really want to carry out the command. This was added because
once the aliases are removed, there is no going back (unless you add
them all back manually or if you have a backup of course).