Modify _specific_commands_help method
When attempting to use the "help" command within a group (ex. "help
<group name> <command>"), the "additional help" of the group was
displayed instead of the specific command(s). This was because the
"_specific_commands_help" method did not have any way of checking within
a group and so only displayed information about the group iself. The
method has been modified and will now display information about the
command(s) within the group.