Add special hanlding for certain aliased commands
If an alias of the "help" or "clear" command was called, nothing
occured because there was no specific handling present for those two
built-in commands (the "alias" command worked fine because handling
already existed for it). To remedy this issue, the method
"_special_alias_builtins" was created to handle aliases of the "help"
and "clear" commands. Additionally, a check was also added in the
"magnetsting_mainloop" method to determine if a command falls within the
special category.