Remove unnecessary alias file write
When using the "alias remove *" command, the alias file was
unnecessarily written to if the user had cancelled the deletions. The
contents of the file were not altered, rather, it just wrote the entire
alias dict to the file again. This was due to the small part of code
responsible for handling the writing being in the wrong spot. It has
since been moved to the correct position and will only update the alias
file if the user confirms removal.