Reserve "*" as a special keyword for aliases
The "*" character has been reserved as a special keyword and cannot be
used an an alias name. This was done because if someone tried to remove
the alias, they would be unable to as "alias remove *" would result in
the removal of all aliases, not just the alias named "*". This change
also helps prevent the accidental removal of all aliases at once.