Rewrite long_help of help command
The "long_help" of the "help" command has been rewritten to be more
informative of what can be done using the command. This was done
because the previous description was rather lacking in information on
how the "help" command can be used beyond displaying the main help
banner. This new incarnation now also includes information on how
"help" can be used to search for and provide additional help for
specific commands and groups and how it can be used within command
groups as well.