Edit example call of file-type command in README
The name of the file-type command example call has been changed to
match the name of the file-type command used in the file-type command
creation example. This was done to avoid any confusion as the other
example calls all used the same command names as was used in the
creation examples.