Specify exception raised for non-existent groups
The README file has been edited to specify the name of the exception
raised when attempting to assign a command to a non-existent group.
This was done as the README used to only say that an exception was
raised, but did not mention what the exact exception was called. This
will help both with giving users a "heads up" of the exception that
is raised in those types of circumstances as well as with error
handling.